Add functionality to store and retrieve meta information on ranks
For streaming applications and their performance, it will often be
crucial to have more detailed information on the MPI ranks of an
applications, such as e.g. their hostname.

This commit adds functionality to set/retrieve such information as a
vector of strings with user-defined interpretation and store it as
openPMD attribute.

File reading and collecting such strings across multiple ranks are added
as a convenience.