Error handling at Series level
This is the commit with the most complicated logic changes.
It implements skipping broken iterations at the Series level.
This includes the Streaming API (Series::readIterations()), as well as
our three iteration encodings.

On the positive side: The most complex logic changes have already been
prepared in the topic-adios2-append PR